Update run-move-locally.md to work with the current Libra system
After https://github.com/libra/libra/pull/4820, the LibraAccount APIs
for withdraw_from and deposit_to are not available, so the simple identity
function used as an example in MyModule.move cannot be used with Libra
coin tokens. It wasn't a very good example, anyway. I've replaced it here
by a slightly more complex example that adds a processing fee to a payment.
I used 2 custom scripts: one to specify the fee amount and another to send
a payment with the additional fee.